fix: 게시글 삭제 서비스에 예외처리 추가

+++ src/main/java/com/sammaru5/sammaru/service/article/ArticleRemoveService.java
@@ -42,6 +42,10 @@ public class ArticleRemoveService {
     }
 
     public boolean removeArticleByAdmin(Long boardId) {
+        if (!articleRepository.existsById(boardId)) {
+            throw new CustomException(ErrorCode.BOARD_NOT_FOUND, boardId.toString());
+        }
+
         articleRepository.deleteArticlesByBoardId(boardId);
         return true;
     }
